+++ src/Router.tsx
@@ -4,10 +4,10 @@ import Landing from './pages/Landing'
 import Login from './pages/Login'
 import KakaoRedirection from './pages/KakaoRedirection'
 import Spotify from './pages/Spotify'
-import SignUp from './pages/SignUp'
+import SignUpTerms from './pages/SignUpTerms'
+import SignUpNickname from './pages/SignUpNickname'
 import Setting from './pages/Setting'
 import EditProfile from './pages/EditProfile'
-import SignUpTerms from './pages/SignUpTerms'
 
 const Router = () => {
   return (
@@ -16,9 +16,9 @@ const Router = () => {
         <Route path="/" element={<Outer />}>
           <Route path="" element={<Landing />} />
           <Route path="login" element={<Login />} />
-          <Route path="signup/terms" element={<SignUpTerms />} />
           <Route path="auth" element={<KakaoRedirection />} />
-          <Route path="signup" element={<SignUp />} />
+          <Route path="signup/terms" element={<SignUpTerms />} />
+          <Route path="signup/nickname" element={<SignUpNickname />} />
           <Route path="spotify" element={<Spotify />} />
           <Route path="setting" element={<Setting />} />
           <Route path="editProfile" element={<EditProfile />} />
